pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lora-osm-notes-gateway"
version = "0.1.0"
description = "Gateway Meshtastic USB → OSM Notes for Raspberry Pi"
requires-python = ">=3.8"
dependencies = [
    "pyserial>=3.5",
    "protobuf>=4.25.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
